requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
requests>=2.28.0
plotly>=5.18.0
fpdf2>=2.7.0
